Finished Skeleton
Added in admin funcationality for an administrator to create and account to be able to create/edit products that will be shown on the home page. also created the cart functionality which whill create a cart with a unqiue ID that is tracked within the customers cookies to remmeber what they have put in their cart.